seperated modals and views into their own folders and files so now bot code is a lot easier to navigate, which means i can actually think when coding!! also fixed a few bugs around the setup command
