+++ modals/__init__.py
@@ -1,0 +1,6 @@
+from modals.denyhours import DenyHours
+from modals.edithours import EditHours
+from modals.setsheetname import SetSheetName
+from modals.addsubteam import AddSubteam
+
+__all__ = ["DenyHours", "EditHours", "SetSheetName", "AddSubteam"]
