+++ views/__init__.py
@@ -1,0 +1,5 @@
+from views.verifyhours import VerifyHours
+from views.setuppages import SetupPages
+from views.plusminus import PlusMinus
+
+__all__ = ["VerifyHours", "SetupPages", "PlusMinus"]
